Impolements multiple changes to the project structure from Sergii's PR review.
