+++ src/main/java/echo/App.java
@@ -1,4 +1,4 @@
-package org.example;
+package echo;
 
 import java.io.IOException;
 import java.net.ServerSocket;
